requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
APScheduler==3.9.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
Django==4.0.4
frozenlist==1.2.0
future==0.18.2
idna==3.3
jieba==0.42.1
jsonschema==3.2.0
line-bot-sdk==2.0.1
multidict==5.2.0
pycparser==2.21
pyrsistent==0.18.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
yarl==1.7.2
